+++ patch_api/api/estimates_api.py
@@ -35,8 +35,11 @@ class EstimatesApi(object):
         "page",
         "distance_m",
         "transportation_method",
-        "package_weight_g",
+        "package_mass_g",
         "create_order",
+        "model",
+        "make",
+        "year",
     ]
 
     def __init__(self, api_client=None):
@@ -111,8 +114,11 @@ class EstimatesApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -251,8 +257,11 @@ class EstimatesApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -304,6 +313,292 @@ class EstimatesApi(object):
 
         return self.api_client.call_api(
             "/v1/estimates/mass",
+            "POST",
+            path_params,
+            query_params,
+            header_params,
+            body=body_params,
+            post_params=form_params,
+            files=local_var_files,
+            response_type="EstimateResponse",  # noqa: E501
+            auth_settings=auth_settings,
+            async_req=local_var_params.get("async_req"),
+            _return_http_data_only=local_var_params.get(
+                "_return_http_data_only"
+            ),  # noqa: E501
+            _preload_content=local_var_params.get("_preload_content", True),
+            _request_timeout=local_var_params.get("_request_timeout"),
+            collection_formats=collection_formats,
+        )
+
+    def create_shipping_estimate(
+        self, create_shipping_estimate_request={}, **kwargs
+    ):  # noqa: E501
+        """Create a shipping estimate given the distance traveled in meters, package weight, and transportation method.  # noqa: E501
+
+        Creates a shipping estimate for the amount of CO2 to be compensated. An order in the `draft` state may be created based on the parameters.   # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+        >>> thread = api.create_shipping_estimate(create_shipping_estimate_request, async_req=True)
+        >>> result = thread.get()
+
+        :param async_req bool: execute request asynchronously
+        :param CreateShippingEstimateRequest create_shipping_estimate_request: (required)
+        :param _preload_content: if False, the urllib3.HTTPResponse object will
+                                 be returned without reading/decoding response
+                                 data. Default is True.
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :return: EstimateResponse
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+        kwargs["_return_http_data_only"] = True
+        return self.create_shipping_estimate_with_http_info(
+            create_shipping_estimate_request, **kwargs
+        )  # noqa: E501
+
+    def create_shipping_estimate_with_http_info(
+        self, create_shipping_estimate_request, **kwargs
+    ):  # noqa: E501
+        """Create a shipping estimate given the distance traveled in meters, package weight, and transportation method.  # noqa: E501
+
+        Creates a shipping estimate for the amount of CO2 to be compensated. An order in the `draft` state may be created based on the parameters.   # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+        >>> thread = api.create_shipping_estimate_with_http_info(create_shipping_estimate_request, async_req=True)
+        >>> result = thread.get()
+
+        :param async_req bool: execute request asynchronously
+        :param CreateShippingEstimateRequest create_shipping_estimate_request: (required)
+        :param _return_http_data_only: response data without head status code
+                                       and headers
+        :param _preload_content: if False, the urllib3.HTTPResponse object will
+                                 be returned without reading/decoding response
+                                 data. Default is True.
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :return: tuple(EstimateResponse, status_code(int), headers(HTTPHeaderDict))
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+
+        local_var_params = locals()
+
+        all_params = ["create_shipping_estimate_request"]  # noqa: E501
+        all_params.append("async_req")
+        all_params.append("_return_http_data_only")
+        all_params.append("_preload_content")
+        all_params.append("_request_timeout")
+        all_params.append("mass_g")
+        all_params.append("price_cents_usd")
+        all_params.append("project_id")
+        all_params.append("metadata")
+        all_params.append("distance_m")
+        all_params.append("transportation_method")
+        all_params.append("package_mass_g")
+        all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
+
+        for key, val in six.iteritems(local_var_params["kwargs"]):
+            if key not in all_params:
+                raise ApiTypeError(
+                    "Got an unexpected keyword argument '%s'"
+                    " to method create_shipping_estimate" % key
+                )
+            local_var_params[key] = val
+        del local_var_params["kwargs"]
+        # verify the required parameter 'create_shipping_estimate_request' is set
+        if (
+            "create_shipping_estimate_request" not in local_var_params
+            or local_var_params["create_shipping_estimate_request"] is None
+        ):
+            raise ApiValueError(
+                "Missing the required parameter `create_shipping_estimate_request` when calling `create_shipping_estimate`"
+            )  # noqa: E501
+
+        collection_formats = {}
+
+        path_params = {}
+
+        query_params = []
+        for key in kwargs:
+            query_params.append([key, kwargs.get(key)])
+
+        header_params = {}
+
+        form_params = []
+        local_var_files = {}
+
+        body_params = None
+        if "create_shipping_estimate_request" in local_var_params:
+            body_params = local_var_params["create_shipping_estimate_request"]
+        # HTTP header `Accept`
+        header_params["Accept"] = self.api_client.select_header_accept(
+            ["application/json"]
+        )  # noqa: E501
+
+        # HTTP header `Content-Type`
+        header_params[
+            "Content-Type"
+        ] = self.api_client.select_header_content_type(  # noqa: E501
+            ["application/json"]
+        )  # noqa: E501
+
+        # Authentication setting
+        auth_settings = ["bearer_auth"]  # noqa: E501
+
+        return self.api_client.call_api(
+            "/v1/estimates/shipping",
+            "POST",
+            path_params,
+            query_params,
+            header_params,
+            body=body_params,
+            post_params=form_params,
+            files=local_var_files,
+            response_type="EstimateResponse",  # noqa: E501
+            auth_settings=auth_settings,
+            async_req=local_var_params.get("async_req"),
+            _return_http_data_only=local_var_params.get(
+                "_return_http_data_only"
+            ),  # noqa: E501
+            _preload_content=local_var_params.get("_preload_content", True),
+            _request_timeout=local_var_params.get("_request_timeout"),
+            collection_formats=collection_formats,
+        )
+
+    def create_vehicle_estimate(
+        self, create_vehicle_estimate_request={}, **kwargs
+    ):  # noqa: E501
+        """Create a vehicle estimate given the distance traveled in meters and the type of vehicle  # noqa: E501
+
+        Creates an estimate and calculates the amount of CO2 to be compensated depending on the distance and the vehicle. An order in the `draft` state may be created based on the parameters, linked to the estimate.   # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+        >>> thread = api.create_vehicle_estimate(create_vehicle_estimate_request, async_req=True)
+        >>> result = thread.get()
+
+        :param async_req bool: execute request asynchronously
+        :param CreateVehicleEstimateRequest create_vehicle_estimate_request: (required)
+        :param _preload_content: if False, the urllib3.HTTPResponse object will
+                                 be returned without reading/decoding response
+                                 data. Default is True.
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :return: EstimateResponse
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+        kwargs["_return_http_data_only"] = True
+        return self.create_vehicle_estimate_with_http_info(
+            create_vehicle_estimate_request, **kwargs
+        )  # noqa: E501
+
+    def create_vehicle_estimate_with_http_info(
+        self, create_vehicle_estimate_request, **kwargs
+    ):  # noqa: E501
+        """Create a vehicle estimate given the distance traveled in meters and the type of vehicle  # noqa: E501
+
+        Creates an estimate and calculates the amount of CO2 to be compensated depending on the distance and the vehicle. An order in the `draft` state may be created based on the parameters, linked to the estimate.   # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+        >>> thread = api.create_vehicle_estimate_with_http_info(create_vehicle_estimate_request, async_req=True)
+        >>> result = thread.get()
+
+        :param async_req bool: execute request asynchronously
+        :param CreateVehicleEstimateRequest create_vehicle_estimate_request: (required)
+        :param _return_http_data_only: response data without head status code
+                                       and headers
+        :param _preload_content: if False, the urllib3.HTTPResponse object will
+                                 be returned without reading/decoding response
+                                 data. Default is True.
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :return: tuple(EstimateResponse, status_code(int), headers(HTTPHeaderDict))
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+
+        local_var_params = locals()
+
+        all_params = ["create_vehicle_estimate_request"]  # noqa: E501
+        all_params.append("async_req")
+        all_params.append("_return_http_data_only")
+        all_params.append("_preload_content")
+        all_params.append("_request_timeout")
+        all_params.append("mass_g")
+        all_params.append("price_cents_usd")
+        all_params.append("project_id")
+        all_params.append("metadata")
+        all_params.append("distance_m")
+        all_params.append("transportation_method")
+        all_params.append("package_mass_g")
+        all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
+
+        for key, val in six.iteritems(local_var_params["kwargs"]):
+            if key not in all_params:
+                raise ApiTypeError(
+                    "Got an unexpected keyword argument '%s'"
+                    " to method create_vehicle_estimate" % key
+                )
+            local_var_params[key] = val
+        del local_var_params["kwargs"]
+        # verify the required parameter 'create_vehicle_estimate_request' is set
+        if (
+            "create_vehicle_estimate_request" not in local_var_params
+            or local_var_params["create_vehicle_estimate_request"] is None
+        ):
+            raise ApiValueError(
+                "Missing the required parameter `create_vehicle_estimate_request` when calling `create_vehicle_estimate`"
+            )  # noqa: E501
+
+        collection_formats = {}
+
+        path_params = {}
+
+        query_params = []
+        for key in kwargs:
+            query_params.append([key, kwargs.get(key)])
+
+        header_params = {}
+
+        form_params = []
+        local_var_files = {}
+
+        body_params = None
+        if "create_vehicle_estimate_request" in local_var_params:
+            body_params = local_var_params["create_vehicle_estimate_request"]
+        # HTTP header `Accept`
+        header_params["Accept"] = self.api_client.select_header_accept(
+            ["application/json"]
+        )  # noqa: E501
+
+        # HTTP header `Content-Type`
+        header_params[
+            "Content-Type"
+        ] = self.api_client.select_header_content_type(  # noqa: E501
+            ["application/json"]
+        )  # noqa: E501
+
+        # Authentication setting
+        auth_settings = ["bearer_auth"]  # noqa: E501
+
+        return self.api_client.call_api(
+            "/v1/estimates/vehicle",
             "POST",
             path_params,
             query_params,
@@ -385,8 +680,11 @@ class EstimatesApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -509,8 +807,11 @@ class EstimatesApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params: