+++ patch_api/api/orders_api.py
@@ -35,8 +35,11 @@ class OrdersApi(object):
         "page",
         "distance_m",
         "transportation_method",
-        "package_weight_g",
+        "package_mass_g",
         "create_order",
+        "model",
+        "make",
+        "year",
     ]
 
     def __init__(self, api_client=None):
@@ -105,8 +108,11 @@ class OrdersApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -231,8 +237,11 @@ class OrdersApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -365,8 +374,11 @@ class OrdersApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -489,8 +501,11 @@ class OrdersApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -613,8 +628,11 @@ class OrdersApi(object):
         all_params.append("metadata")
         all_params.append("distance_m")
         all_params.append("transportation_method")
-        all_params.append("package_weight_g")
+        all_params.append("package_mass_g")
         all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params: